energy_api/requirements.txt

flask
joblib
numpy
pandas
scikit-learn
xgboost
catboost
lightgbm
python-dotenv
gunicorn
flask-cors
